Rewrite everything using T3 and electron
 * Now using T3 to structure the app
 * Switched from NW.js to electron
 * Completely changed the way projects are handled:
  * All projects are now created in [user folder]/toothrot/projects/
  * Remove manually selecting things in the file system

All of these changes allow us to transforming toothrot-builder into
a real editor instead of just a GUI for toothrot's API.

Please note that it was necessary to selectively apply monkey patching
for the fs module. The reason for this is that the electron developers
decided to extend the original node fs module with the ability to
treat ASAR files as directories, so that the fs module in electron
apps behaves slightly different than the original. Both rimraf and
the asar module itself cannot be used in an electron app without
this kind of patching.